refactor all launch_configuration.go
